Project Content update
I have refined the app, added screenshots of each activity and of the code from MIT app inventor. 

Changed some more formatting of the ReadMe and will again as I need to for screen shots. 

Changed the servo pins for wrist_ver and wrist_rot as they were wrong. Also made the default position the same as the apps sliders #consistency